refactor: extract indexer/infer/sig.rs and args.rs (phase 2)
- Extract 9 pure signature-extraction fns into src/indexer/infer/sig.rs:
  collect_signature, find_fun_signature{,_full}, collect_{fun_params_text,
  all_fun_params_texts,params_from_line}, nth/last_fun_param_type_str,
  strip_trailing_call_args
- Promote find_fun_signature_with_receiver from impl Indexer method to
  pub(crate) free function in sig.rs; update sole call site in backend.rs
- Extract 6 pure call-argument parsing fns into src/indexer/infer/args.rs:
  find_as_call_arg_type, extract_{first_arg,named_arg_name},
  find_named_param_type_in_sig, lambda_param_position_on_line,
  has_named_params_not_it
- Make is_id_char and find_enclosing_call_name pub(crate) in indexer.rs
  so submodules can import them
- Add sig_tests.rs (16 tests) and args_tests.rs (22 tests) as separate
  test files, referenced via #[cfg(test)] #[path = ...] mod tests;
- Re-export all extracted items pub(crate) from indexer.rs so all existing
  callers (inline tests, lambda.rs, backend.rs) continue to resolve by name
- All 393 tests pass (346 pre-existing + 47 new)
